requirements.txt
python-dotenv
graphviz
mistralai
flask